+++ geo/CHANGES.md
@@ -35,6 +35,10 @@
   - Consolidate traits `TryMapCoords` into `MapCoords` and `TryMapCoordsInplace` into `MapCoordsInPlace`
 * Implement `ChamberlainDuquetteArea` for all geo types.
   * <https://github.com/georust/geo/pull/833>
+* Add `{Convert,TryConvert}` traits for coordinate value type conversion.
+  * <https://github.com/georust/geo/pull/836>
+* BREAKING: MapCoords/MapCoordsInPlace now map `Coordinate`s rather than `(x,y)` tuples
+  * <https://github.com/georust/geo/pull/837>
 
 ## 0.20.1